pyproject.toml
[tool.poetry]
name = "truth-extractor"
version = "1.0.0"
description = "Production-ready web crawler for extracting normalized business information"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "truth_extractor"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
requests-cache = "^1.1.1"
tldextract = "^5.1.1"
beautifulsoup4 = "^4.12.2"
lxml = ">=4.9.3"
phonenumbers = ">=8.13.26"
dnspython = "^2.4.2"
Pillow = "^10.1.0"
pydantic = "^2.5.0"
pyyaml = "^6.0.1"
jsonschema = "^4.20.0"
colorama = "^0.4.6"
playwright = {version = "^1.40.0", optional = true}

[tool.poetry.extras]
playwright = ["playwright"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.12.0"
mypy = "^1.7.1"

[tool.poetry.scripts]
truth-extractor = "truth_extractor.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

